refactor: merge load_dataset and dataset
