--- codecov.yml
@@ -1,2 +1,0 @@
-ignore:
-    - src/load_dataset.py
